requirements.txt
paho-mqtt==1.6.1
colorama==0.4.6